requirements.txt
scipy==1.16.2
albumentations==2.0.8
matplotlib==3.10.6
numpy
opencv_python==4.12.0.88
pandas==2.3.3
Pillow==12.0.0
scikit_learn==1.7.2
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1

